LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/websocket-cookie-overwrite-behavior-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/cookie?set' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 96 - No message available on STREAM)
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/cookie?set' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 96 - No message available on STREAM)
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/cookie?set' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 96 - No message available on STREAM)
Tests WebSocket Set-Cookie overwriting behavior with respect to a document cookie.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Setting third-party cookie 'foo' through cross-origin WebSocket handshake and checking that it doesn't write first-party cookies.
PASS cookieValue is ""

Opening localhost third-party iframe to check its cookies.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Cookies are: setAsFirstParty = value foo = bar
